Update sockets.js
changed intro message! I think the sites users would like this better, as there are lots of complaints about the tone of the current one being a little condescending. The new message: ⚡️please be kind, we were all new once ⚡️ is something that was in a new players bio that we all found to be a cute, fun, and not too self-important reminder that we should all be nice to newer (and older!) players.