Update sockets.js
changed the message that appears on load-in from "This is a gentle reminder that toxic language and abuse is not tolerated on ProAvalon. Please report instances of rule violations to a moderator. If you are unhappy with a moderator's decision, please contact ProNub via discord or email to review the case." to "⚡️ please be kind, we were all new once ⚡️".